requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-cloud-speech==2.21.0
google-auth==2.23.4
python-multipart==0.0.6
pydantic==2.5.0